chore: moved model out of firebase
