+++ app/src/main/java/com/simplecity/amp_library/ui/screens/nowplaying/PlayerFragment.java
@@ -77,7 +77,7 @@ import com.simplecity.amp_library.utils.StringUtils;
 import com.simplecity.amp_library.utils.color.ArgbEvaluator;
 import com.simplecity.amp_library.utils.menu.song.SongMenuUtils;
 import dagger.android.support.AndroidSupportInjection;
-import edu.usf.sas.pal.muser.firebase.model.EventProvider;
+import edu.usf.sas.pal.muser.model.EventProvider;
 import edu.usf.sas.pal.muser.util.EventUtils;
 import io.reactivex.BackpressureStrategy;
 import io.reactivex.Flowable;
