+++ app/src/main/java/edu/usf/sas/pal/muser/model/SongData.kt
@@ -1,4 +1,4 @@
-package edu.usf.sas.pal.muser.firebase.model
+package edu.usf.sas.pal.muser.model
 
 
 /**
